Trying to push the code of assign1
